pytorch

+++ tests/beignet/_polynomial/test_hermline.py
@@ -6,5 +6,5 @@ import torch
 def test_hermline():
     torch.testing.assert_close(
         beignet.polynomial._hermline.hermline(3, 4),
-        torch.tensor([3, 2]),
+        torch.tensor([3, 2], dtype=torch.float32),
     )
